Update index.md
I edited the section about the expanded curricula. I think I understand correctly what the paragraph is trying to say. But I'm not sure. 

My understanding is that you can elect to have a standard Library Carpentry workshop (as opposed to a Library Carpentry-based workshop] you must choose one of the four curricula from the table.

I think my confusion is around the use of the word "standard" to describe a workshop and also as a keyword in the title of one of the options for the curriculum. 

Apologies if my edits are too fussy.