Remove Atom (retired) from the proposed editors
While looking over the course material, I clicked the link to Atom since I was curious about it as an alternative to VS Code. I dicovered this editor has been sunset to be able to focus efforts on VS Code. I therefore suggest removing it from the suggested editors and adjusting the respective sentence accordingly.